+++ assets/javascripts/kitten/components/form/text-input-with-unit/index.js
@@ -125,6 +125,7 @@ export const TextInputWithUnit = ({
         {...others}
         size={size}
         className={classNames('k-Form-TextInputWithUnit__input', className)}
+        variant={variant}
       />
       <span
         className={classNames(